+++ src/main/java/com/controlcenter/controlcenter/controller/DevPhaseController.java
@@ -1,5 +1,6 @@
 package com.controlcenter.controlcenter.controller;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
@@ -37,25 +38,25 @@ public class DevPhaseController {
     private ErrorHandler errorHandler;
 
     @GetMapping("/all")
-    public ResponseEntity<List<DevPhaseOutput>> getAllDevPhase() {
-         // Check if the user is authenticated 
-        //  Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
+    public ResponseEntity<List<DevPhaseOutput>> getAllDevPhase(HttpSession httpSession) {
+        // Check if the user is authenticated 
+         Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
         
-        //  if (isAuthenticated != null && isAuthenticated) {
+         if (isAuthenticated != null && isAuthenticated) {
              // User is authenticated  
             return devPhaseService.getAllDevPhase();
-        // } else {
-        //     // User is not authenticated
-        //     return ResponseEntity.ok(new ArrayList<DevPhaseOutput>());
-        // }
+        } else {
+            // User is not authenticated
+            return ResponseEntity.ok(new ArrayList<DevPhaseOutput>());
+        }
     }
 
     @PostMapping("/add")
-    public ResponseEntity<String> addDevPhase(@RequestBody DevPhaseInput devPhase){
+    public ResponseEntity<String> addDevPhase(@RequestBody DevPhaseInput devPhase, HttpSession httpSession){
         // Check if the user is authenticated
-        // Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
+        Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
 
-        // if (isAuthenticated != null && isAuthenticated) {
+        if (isAuthenticated != null && isAuthenticated) {
             // User is authenticated,  proceed with adding
             //For Validation
             ValidatorFactory validatorFactory = Validation.buildDefaultValidatorFactory();
@@ -68,18 +69,18 @@ public class DevPhaseController {
                     String emp_id = "101"; //httpSession.getAttribute("session").toString();
                     return ResponseEntity.status(200).body(devPhaseService.addDevPhase(devPhase, emp_id));
                 }
-        // } else {
-        //     // is not authenticated
-        //     return ResponseEntity.status(401).body("Unauthorized");
-        // }
+        } else {
+            // is not authenticated
+            return ResponseEntity.status(401).body("Unauthorized");
+        }
     }
 
     @PutMapping("/edit/{id}")
-    public ResponseEntity<String> editDevPhaseInfo(@PathVariable String id, @RequestBody DevPhaseInput devPhase) {
+    public ResponseEntity<String> editDevPhaseInfo(@PathVariable String id, @RequestBody DevPhaseInput devPhase, HttpSession httpSession) {
         // Check if the user is authenticated
-        // Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
+        Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
         
-        // if (isAuthenticated != null && isAuthenticated){
+        if (isAuthenticated != null && isAuthenticated){
             // User is authenticated,  proceed with adding
             //For Validation
             ValidatorFactory validatorFactory = Validation.buildDefaultValidatorFactory();
@@ -92,61 +93,61 @@ public class DevPhaseController {
                     String emp_id = "101"; //httpSession.getAttribute("session").toString();
                     return devPhaseService.editDevPhaseInfo(id, devPhase, emp_id);
                 }
-        // } else {
-        //     // User is not authenticated 
-        //     return ResponseEntity.status(401).body("Unauthorized");
-        // }
+        } else {
+            // User is not authenticated 
+            return ResponseEntity.status(401).body("Unauthorized");
+        }
     }
 
     @PutMapping("/delete-multiple")
-    public ResponseEntity<String> deleteMultipleDevPhase(@RequestParam List<Long> ids) {
+    public ResponseEntity<String> deleteMultipleDevPhase(@RequestParam List<Long> ids, HttpSession httpSession) {
         // Check uf the user is authenticated
-        // Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
+        Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
 
-        // if (isAuthenticated != null && isAuthenticated){
+        if (isAuthenticated != null && isAuthenticated){
             try {
                 String emp_id = "101"; //httpSession.getAttribute("session").toString();
             return ResponseEntity.ok().body(devPhaseService.deleteMultipleDevPhase(ids, emp_id));
             } catch (Exception e) {
                 return ResponseEntity.status(500).body("Server Side Error");
             }
-        // } else {
-        //     // User is not authenticated
-        //     return ResponseEntity.status(401).body("Unauthorized");
-        // }
+        } else {
+            // User is not authenticated
+            return ResponseEntity.status(401).body("Unauthorized");
+        }
     }
 
     @PutMapping("/delete/{id}")
-    public ResponseEntity<String> logicalDeleteDevPhase(@PathVariable String id) {
+    public ResponseEntity<String> logicalDeleteDevPhase(@PathVariable String id, HttpSession httpSession) {
         // Check if the user is authenticated
-        // Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
+        Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
 
-        // if (isAuthenticated != null && isAuthenticated){
+        if (isAuthenticated != null && isAuthenticated){
             try {
                 String emp_id = "101"; //httpSession.getAttribute("session").toString();
                 return ResponseEntity.ok().body(devPhaseService.logicalDeleteDevPhase(id, emp_id));
             } catch (Exception e) {
                 return ResponseEntity.status(500).body("Server Side Error");
             }
-        // } else {
-        //     // User is not authenticated
-        //     return ResponseEntity.status(401).body("Unauthorized");
-        // }
+        } else {
+            // User is not authenticated
+            return ResponseEntity.status(401).body("Unauthorized");
+        }
     }
 
     @PutMapping("/restore/{id}")
     public ResponseEntity<String> restoreDevPhase(@PathVariable String id, HttpSession httpSession) {
-        // Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
+        Boolean isAuthenticated = (Boolean) httpSession.getAttribute("isAuthenticated");
 
-        // if (isAuthenticated != null && isAuthenticated){
+        if (isAuthenticated != null && isAuthenticated){
             try {
                 return ResponseEntity.ok().body(devPhaseService.restoreDevPhase(id));
             } catch (Exception e) {
                 return ResponseEntity.status(500).body("Server Side Error");
             }
-        // } else {
-        // // User is not authenticated
-        // return ResponseEntity.status(401).body("Unauthorized");
-        // }  
+        } else {
+        // User is not authenticated
+        return ResponseEntity.status(401).body("Unauthorized");
+        }  
     }
 }
